+++ iati/default.py
@@ -123,7 +123,7 @@ def _codelists(version, use_cache=False):
     """
     version = _specific_version_for(version)
 
-    paths = iati.resources.get_all_codelist_paths(version)
+    paths = iati.resources.get_codelist_paths(version)
 
     for path in paths:
         _, filename = os.path.split(path)
@@ -170,7 +170,7 @@ def codelist_mapping(version):
     """
     version = _specific_version_for(version)
 
-    path = iati.resources.get_codelist_mapping_path(version)
+    path = iati.resources.create_codelist_mapping_path(version)
     mapping_tree = iati.utilities.load_as_tree(path)
     mappings = defaultdict(list)
 
@@ -206,7 +206,7 @@ def ruleset(version):
     """
     version = _specific_version_for(version)
 
-    path = iati.resources.get_ruleset_path(iati.resources.FILE_RULESET_STANDARD_NAME, version)
+    path = iati.resources.create_ruleset_path(iati.resources.FILE_RULESET_STANDARD_NAME, version)
     ruleset_str = iati.utilities.load_as_string(path)
 
     return iati.Ruleset(ruleset_str)
@@ -225,7 +225,7 @@ def ruleset_schema():
         Determine whether a version should be provided. This is worth considering if the content of the IATI Ruleset Schema varies between versions.
 
     """
-    path = iati.resources.get_ruleset_path(iati.resources.FILE_RULESET_SCHEMA_NAME)
+    path = iati.resources.create_ruleset_path(iati.resources.FILE_RULESET_SCHEMA_NAME)
     schema_str = iati.utilities.load_as_string(path)
 
     return json.loads(schema_str)
@@ -332,7 +332,7 @@ def activity_schema(version, populate=True):
         iati.ActivitySchema: An instantiated IATI Schema for the specified version of the Standard.
 
     """
-    return _schema(iati.resources.get_all_activity_schema_paths, iati.ActivitySchema, version, populate)
+    return _schema(iati.resources.get_activity_schema_paths, iati.ActivitySchema, version, populate)
 
 
 def organisation_schema(version, populate=True):
@@ -349,4 +349,4 @@ def organisation_schema(version, populate=True):
         iati.OrganisationSchema: An instantiated IATI Schema for the specified version of the Standard.
 
     """
-    return _schema(iati.resources.get_all_organisation_schema_paths, iati.OrganisationSchema, version, populate)
+    return _schema(iati.resources.get_organisation_schema_paths, iati.OrganisationSchema, version, populate)